ALU.circ
Design a 3-bit ALU using Logisim. Each bit cell should be able to do
ADD, AND, NOT and OR.

In addition, the adder should use Carry-Lookahead to improve
performance. Your 3-bit ALU should employ group carry lookahead, with
group size of 3 and ripple carry between groups. (In other words,
correctly generate the carry out of the first group).

You are to assume that all arithmetic operations are on 2's complement
numbers. The NOT function should occur on the A inputs. The
selector/control bits: AND=00, OR=01, NOT=10, ADD=11.